paranın db'ye yazılma işlemi eklendi

+++ Clicker.Entities/Concrete/Improvement.cs
@@ -11,7 +11,7 @@ namespace Clicker.Entities.Concrete
     {
         public string Id { get; set; }
         public double CurrentPrice { get; set; }
-        public int TimeMs { get; set; }
+        public double TimeMs { get; set; }
         public int UpgradeCount { get; set; }
         public bool Manager { get; set; }
         public double DefaultMoney { get; set; }
